backend/reproduce_error.py
import requests
import sys

BASE_URL = "http://localhost:8000/api/v1"

def reproduce():
    # 1. Login to get token (assuming test user exists from previous runs or seed)
    # If not, we might need to signup first. Let's try to signup a new user to be sure.
    email = "debug_user_1@example.com"
    password = "Password123!"
    
    # Signup
    requests.post(f"{BASE_URL}/auth/signup", json={
        "email": email,
        "password": password,
        "confirm_password": password,
        "full_name": "Debug User"
    })
    
    # We need to be a developer.
    # We can't easily promote via API without admin token.
    # But wait, in the previous `test_create_app.py`, it promoted via DB.
    # I can't check DB easily without python async script.
    
    # However, the user said "Request failed with status code 500".
    # This implies they successfully authenticated and reached the endpoint.
    # If I get 403, I know it's permissions.
    
    # Let's try to login as the user from `test_create_app.py` if it exists?
    # Or just try to hit the endpoint unauthenticated and see if it is 401.
    
    # STARTUP PROBLEM: I cannot easily promote a user to developer against a running server without stopping it to run a script, or using a separate script that connects to DB.
    # BUT, I can run a script that connects to DB, promotes the user, and then hits the API.
    pass

if __name__ == "__main__":
    # I will rely on the `test_create_app.py` structure which handles promotion.
    pass
