backend/requirements.txt
# Backend Dependencies
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP Client (for MCP)
httpx==0.26.0
aiohttp==3.9.1

# AI (Gemini)
google-generativeai>=0.4.0

# Security
cryptography==41.0.7

# Validation
email-validator==2.1.0.post1

# Logging & Monitoring
structlog==24.1.0
psutil==5.9.8

# Utilities
python-dotenv==1.0.0
orjson==3.9.12

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
